pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ievo-sdk"
version = "0.1.0"
description = "Developer toolkit for building iEvo agents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "iEvo", email = "hello@ievo.ai" }]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "jsonschema>=4.20.0",
    "jinja2>=3.1.0",
]

[project.scripts]
ievo-sdk = "ievo_sdk.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8.0"]

[tool.uv]
dev-dependencies = ["pytest>=8.0", "ruff>=0.8.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/ievo_sdk"]

[tool.ruff]
target-version = "py310"
line-length = 100
